add intermediary `run.js` script
This shouldn't be necessary, but `babel-cli` ends up
stripping the she bang from the input file, which makes
the output file not be properly executable.

This little workaround avoids that, by having `run.js` be
executable ES5 code, and having it require the compiled
server.js file as its only task.